refactor(structure): unite extract_f0_rmvpe and extract_f0_print into one
