+++ infer/modules/train/extract/extract_f0_print.py
@@ -132,7 +132,87 @@ class FeatureInput(object):
                 except:
                     printt("f0fail-%s-%s-%s" % (idx, inp_path, traceback.format_exc()))
 
-def extract_features(expected_dir: str, cores: int, method_f0: str):
+class FeatureInputRMVPE(object):
+    def __init__(self, is_half: bool, samplerate=16000, hop_size=160):
+        self.is_half = is_half
+        self.fs = samplerate
+        self.hop = hop_size
+
+        self.f0_bin = 256
+        self.f0_max = 1100.0
+        self.f0_min = 50.0
+        self.f0_mel_min = 1127 * np.log(1 + self.f0_min / 700)
+        self.f0_mel_max = 1127 * np.log(1 + self.f0_max / 700)
+
+    def compute_f0(self, path, f0_method):
+        x = load_audio(path, self.fs)
+        # p_len = x.shape[0] // self.hop
+        if f0_method == "rmvpe":
+            if hasattr(self, "model_rmvpe") == False:
+                from rvc.f0.rmvpe import RMVPE
+
+                print("Loading rmvpe model")
+                self.model_rmvpe = RMVPE(
+                    "assets/rmvpe/rmvpe.pt", is_half=self.is_half, device="cuda"
+                )
+            f0 = self.model_rmvpe.compute_f0(x, filter_radius=0.03)
+        return f0
+
+    def coarse_f0(self, f0):
+        f0_mel = 1127 * np.log(1 + f0 / 700)
+        f0_mel[f0_mel > 0] = (f0_mel[f0_mel > 0] - self.f0_mel_min) * (
+            self.f0_bin - 2
+        ) / (self.f0_mel_max - self.f0_mel_min) + 1
+
+        # use 0 or 1
+        f0_mel[f0_mel <= 1] = 1
+        f0_mel[f0_mel > self.f0_bin - 1] = self.f0_bin - 1
+        f0_coarse = np.rint(f0_mel).astype(int)
+        assert f0_coarse.max() <= 255 and f0_coarse.min() >= 1, (
+            f0_coarse.max(),
+            f0_coarse.min(),
+        )
+        return f0_coarse
+
+    def go(self, paths, f0_method):
+        if len(paths) == 0:
+            printt("no-f0-todo")
+        else:
+            printt("todo-f0-%s" % len(paths))
+            n = max(len(paths) // 5, 1)  # 每个进程最多打印5条
+            for idx, (inp_path, opt_path1, opt_path2) in enumerate(paths):
+                try:
+                    if idx % n == 0:
+                        printt("f0ing,now-%s,all-%s,-%s" % (idx, len(paths), inp_path))
+                    if (
+                        os.path.exists(opt_path1 + ".npy") == True
+                        and os.path.exists(opt_path2 + ".npy") == True
+                    ):
+                        continue
+                    featur_pit = self.compute_f0(inp_path, f0_method)
+                    np.save(
+                        opt_path2,
+                        featur_pit,
+                        allow_pickle=False,
+                    )  # nsf
+                    coarse_pit = self.coarse_f0(featur_pit)
+                    np.save(
+                        opt_path1,
+                        coarse_pit,
+                        allow_pickle=False,
+                    )  # ori
+                except:
+                    printt("f0fail-%s-%s-%s" % (idx, inp_path, traceback.format_exc()))
+
+def default_extract_features(expected_dir: str, cores: int, method_f0: str) -> None:
+    """
+    Extract features
+
+    Args:
+        expected_dir: str - Directory to store extracted features
+        cores: int - Number of CPU cores to use
+        method_f0: str - F0 method to use { pm, harvest, dio, rmvpe }
+    """
     featureInput = FeatureInput()
 
     inp_root = f"{expected_dir}/1_16k_wavs" # TODO: rename these vars to not be confusing
@@ -158,3 +238,50 @@ def extract_features(expected_dir: str, cores: int, method_f0: str):
         ps.append(p)
     for p in ps:
         p.join()
+
+def extract_rmvpe_features(expected_dir: str, is_half: bool, i_gpu: int, n_gpu: str):
+    featureInput = FeatureInputRMVPE(is_half=is_half)
+
+    inp_root = f"{expected_dir}/1_16k_wavs" # TODO: rename these vars to not be confusing
+    opt_root1 = f"{expected_dir}/2a_f0"     # TODO: rename these vars to not be confusing
+    opt_root2 = f"{expected_dir}/2b-f0nsf"  # TODO: rename these vars to not be confusing
+    
+    os.makedirs(opt_root1, exist_ok=True)
+    os.makedirs(opt_root2, exist_ok=True)
+
+    paths = []
+
+    for name in sorted(list(os.listdir(inp_root))):
+        inp_path = "%s/%s" % (inp_root, name)
+        if "spec" in inp_path:
+            continue
+        opt_path1 = "%s/%s" % (opt_root1, name)
+        opt_path2 = "%s/%s" % (opt_root2, name)
+        paths.append([inp_path, opt_path1, opt_path2])
+    try:
+        featureInput.go(paths[i_gpu::n_gpu], "rmvpe")
+    except:
+        printt("f0_all_fail-%s" % (traceback.format_exc()))
+
+extraction_methods = {
+           "pm": default_extract_features,
+      "harvest": default_extract_features,
+          "dio": default_extract_features,
+        "rmvpe": default_extract_features,
+}
+
+def extract_features(expected_dir: str, cores: int, method_f0: str) -> None:
+    """
+    Extract features
+
+    Args:
+        expected_dir: str - Directory to store extracted features
+        cores: int - Number of CPU cores to use
+        method_f0: str - F0 method to use { pm, harvest, dio, rmvpe }
+    """
+
+    if method_f0 not in extraction_methods: # Check if the method is valid
+        raise ValueError(f"Unknown feature extraction method: {method_f0}")
+    
+    extraction_methods[method_f0](expected_dir, cores, method_f0)
+
